requirements.txt
pandas
streamlit==1.32.2
azure-core==1.30.1
azure-search-documents==11.6.0b2
tiktoken
langchain-core==0.1.33
langchain-openai==0.1.1
